Refactor: 업로드 네비게이션
