+++ feature/upload/src/main/java/com/acon/feature/upload/UploadRoute.kt
@@ -1,4 +1,4 @@
-package com.acon.acon.navigation.route
+package com.acon.feature.upload
 
 import kotlinx.serialization.Serializable
